Rename ber to ber_publicsearch
So can add a module called ber_closedaccess or similar
to process the closed-access ber dataset which is at
a Small Area level instead of Postcode
