+++ src/drem/download/download.py
@@ -1,3 +1,3 @@
 # flake8: noqa
-from drem.download.ber_public_search import DownloadBERPublicsearch as BERPublicsearch
+from drem.download.ber_publicsearch import DownloadBERPublicsearch as BERPublicsearch
 from drem.download.vo import DownloadValuationOffice as ValuationOffice
